feat(DFP-576): Pending Post feature (#214)
## Description

Closes: 
https://forbole.atlassian.net/browse/DFP-150
https://forbole.atlassian.net/browse/DFP-173
https://forbole.atlassian.net/browse/DFP-540
https://forbole.atlassian.net/browse/DFP-576

This PR adds the pending post feature.

It works by maintaining a `latestPostByUser` and `pendingPost` atom. The
`latestPostByUser` atom is populated with the user's 5 most recent
posts, which are polled automatically whenever the `pendingPost` atom is
not empty.

Whenever the `latestPostByUser` atom changes, the `usePendingPost` hook
will scan the post data, and batch any posts with txHashes that matches
with that of a `pendingPost`, before updating the relevant atoms.

Some areas of note
1. Pending posts cannot be clicked to view post details. This makes
sense as the post is technically not on chain yet, and it would not be
feasible to build an entire fake-post details page to accommodate this.
2. There isn't a good feedback for when a pending post has been properly
broadcasted on chain, as the `transaction success` notification occurs
before BDJuno picks up the post data. I think a possible solution to
this is to send a local toast that says something like `Your post is
live` that redirects the user to the relevant post detail when pressed.
3. All pending posts are placed at the start of the `Posts` array.
4. Pending posts will only appear in the `Discover` tab, as I think it's
a bit strange to show the user their own posts if they're in the
Following tab.
5. The `react-native-toast-notification` package has been extended with
an additional `onPressRetry` handler for toasts. This is because I think
the base `onPress` handler should be reserved for when the user presses
the actual toast, and not the `retry` button. Despite the types being
properly configured, my webstorm IDE still throws a typescript error
though.
6. The `posts` atom has been changed into a `postFamily`. Paired with
the `POST_TYPE` enum, this should make interfacing with the `posts`
recoil state more intuitive and less error prone.

---

### Author Checklist

*All items are required. Please add a note to the item if the item is
not applicable and
please add links to any relevant follow up issues.*

- [x] included the correct [type
prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json)
in the PR title
- [x] provided a link to the relevant issue or specification
- [x] reviewed "Files changed" and left comments if necessary
- [x] confirmed all CI checks have passed

### Reviewers Checklist

*All items are required. Please add a note if the item is not applicable
and please add
your handle next to the items reviewed if you only reviewed selected
items.*

- [ ] confirmed the correct [type
prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json)
in the PR title
- [ ] confirmed all author checklist items have been addressed