fix(DFP-854): Biometrics related password issues (#269)
## Description

Closes: https://forbole.atlassian.net/browse/DFP-854

This PR fixes the following issues related to password manipulation
while biometrics is enabled:

1. If the user has multiple accounts, enabling biometrics only enables
it for the current active account

2. Unlocking with biometrics sometimes produces an incorrect password
error

3. It is not possible to recover the user’s password using only
biometrics.

For 1, the issue was that a derived password for use with biometrics
would only be saved for the active account. This means that if a user
has other accounts on the device, they would not be biometrics-ready as
a derived password would not be saved. The flow has been refactored to
derive a password for all accounts in the form of the `setBiometricData`
function in `SecureStorage/index.tsx`. This function is used in two
events: The first being when the user enables biometrics, and the second
being when the user changes their password while biometrics is on.

For 2, I believe this issue is related to the biometrics passwords being
out of sync due to the app only deriving a password for the active
account when biometrics is enabled. This should be fixed with the
refactor that fixes the issue outlined above.

For 3, I don't believe there is an easy and secure way to allow the user
to recover their password using only biometrics. I think for operations
that require the user's original password (i.e changing accounts,
changing passwords), it would be better to have the user re-input their
passwords manually instead of trying to use biometrics.

The change password flow has also been refactored to fix an issue when
changing passwords while biometrics is enabled. Passwords will now be
changed for every account versus just the active account. This should
also resolve any issues related to changing passwords and swapping
accounts.

---

### Author Checklist

*All items are required. Please add a note to the item if the item is
not applicable and
please add links to any relevant follow up issues.*

- [ ] included the correct [type
prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json)
in the PR title
- [ ] provided a link to the relevant issue or specification
- [ ] reviewed "Files changed" and left comments if necessary
- [ ] confirmed all CI checks have passed

### Reviewers Checklist

*All items are required. Please add a note if the item is not applicable
and please add
your handle next to the items reviewed if you only reviewed selected
items.*

- [ ] confirmed the correct [type
prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json)
in the PR title
- [ ] confirmed all author checklist items have been addressed