feat(DFP-623-DFP-637-DFP-876-DFP-868): Optimistic UI for Relationships (#273)
## Description

Closes:
https://forbole.atlassian.net/browse/DFP-623
https://forbole.atlassian.net/browse/DFP-637
https://forbole.atlassian.net/browse/DFP-755
https://forbole.atlassian.net/browse/DFP-868
https://forbole.atlassian.net/browse/DFP-876

This PR adds optimistic UI and API integration to butter.

Optimistic relationships will be persisted to local storage via MMKV,
and resolved accordingly whenever the user's relationships subscription
is changed.

https://forbole.atlassian.net/browse/DFP-590 is not handled by this PR
as it would be easier to let the optimistic API handle the changing
follow/unfollow count, as handling the counters via a mixture of
optimistic UI and API is not simple arithmetic.

When the user switches profiles, the entire optimistic relationship
state and stored MMKV values are reset. I think this is easier than
persisting the optimistic UI between accounts as in an ideal setting,
all pending transactions should be resolved by the time the user
switches back to the original profile.

Moreover, a `useResetOptimisticUI` has been placed in the
`recoil/optimisticUI` folder. This is non-standard (i.e since it is a
hook, it would seem that it belongs in the hooks/ folder instead), but I
think having it next to the relevant recoil states it affects makes more
sense moving forward.

---

### Author Checklist

*All items are required. Please add a note to the item if the item is
not applicable and
please add links to any relevant follow up issues.*

- [x] included the correct [type
prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json)
in the PR title
- [x] provided a link to the relevant issue or specification
- [x] reviewed "Files changed" and left comments if necessary
- [x] confirmed all CI checks have passed

### Reviewers Checklist

*All items are required. Please add a note if the item is not applicable
and please add
your handle next to the items reviewed if you only reviewed selected
items.*

- [ ] confirmed the correct [type
prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json)
in the PR title
- [ ] confirmed all author checklist items have been addressed