refactor: improve how on-chain authorizations are managed (#412)
## Description

This PR improves the overall management of on-chain permissions (both
feegrant and authz). It's somewhat related to #407, as by making the
overall flow it's going to be easier in the future to understand where
the problem is. However, #407 is **not** closed by this PR. Instead,
that issue should have been solved with
2f4031f09a06aa34486f5f8f4dd36da15538ebce instead.

Closes #415 by adding the missing `MsgCreateReportTypeUrl` to
`RequiredMessageTypesGrant`.
Closes #416 by adding the same missing authorization.

**NOTE**  
Due to some Recoil changes, merging this PR will require all users to
completely wipe and re-initialize the app. This will make sure the
settings are stored properly.

<!-- Add a description of the changes that this PR introduces and the
files that
are the most critical to review. -->

---

### Author Checklist

*All items are required. Please add a note to the item if the item is
not applicable and
please add links to any relevant follow up issues.*

- [ ] included the correct [type
prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json)
in the PR title
- [ ] provided a link to the relevant issue or specification
- [ ] reviewed "Files changed" and left comments if necessary
- [ ] confirmed all CI checks have passed

### Reviewers Checklist

*All items are required. Please add a note if the item is not applicable
and please add
your handle next to the items reviewed if you only reviewed selected
items.*

- [ ] confirmed the correct [type
prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json)
in the PR title
- [ ] confirmed all author checklist items have been addressed