refactor: improve posts data fetching (#497)
## Description

This PR improves how the posts data is fetched from the server, by
adding a new `lastUpdatedDate` field to the `Post` object. This field is
set to the current date time each time a post is fetched from the
server.

This addition should make sure that we trigger the re-rendering of all
the posts-related components (like `PostCard` and its children) each
time the posts list is updated. Consequently, this should re-trigger the
querying of the reactions count which should make it possible to always
have an up-to-date counter for each post.

Closes: #458

<!-- Add a description of the changes that this PR introduces and the
files that
are the most critical to review. -->

---

### Author Checklist

*All items are required. Please add a note to the item if the item is
not applicable and
please add links to any relevant follow up issues.*

- [ ] included the correct [type
prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json)
in the PR title
- [ ] provided a link to the relevant issue or specification
- [ ] reviewed "Files changed" and left comments if necessary
- [ ] confirmed all CI checks have passed

### Reviewers Checklist

*All items are required. Please add a note if the item is not applicable
and please add
your handle next to the items reviewed if you only reviewed selected
items.*

- [ ] confirmed the correct [type
prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json)
in the PR title
- [ ] confirmed all author checklist items have been addressed