feat: improve post details loading (#507)
## Description

Closes [DFP-1165](https://forbole.atlassian.net/browse/DFP-1166)

This PR improves the post details loading page.

The change includes passing an `initialPostData` param into the
`PostDetails` page from the origin page, as the queried post data is
enough to show what is needed in the details page. The `usePostData`
will then receive this data, while also querying the server for a more
up-to-date version, and either display the data from the server, or the
`initialPostData`.

The loading logic of the `PostDetails` page have also been adjusted to
fit this change. The screen will no longer display an initial loading
spinner, instead opting for a lazy loading approach for the reactions,
tips, and comments sections. If this effect is desired, a future
enhancement would be to display a more appealing skeleton loader in
place of this.

---

### Author Checklist

*All items are required. Please add a note to the item if the item is
not applicable and
please add links to any relevant follow up issues.*

- [x] included the correct [type
prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json)
in the PR title
- [x] provided a link to the relevant issue or specification
- [x] reviewed "Files changed" and left comments if necessary
- [x] confirmed all CI checks have passed

### Reviewers Checklist

*All items are required. Please add a note if the item is not applicable
and please add
your handle next to the items reviewed if you only reviewed selected
items.*

- [ ] confirmed the correct [type
prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json)
in the PR title
- [ ] confirmed all author checklist items have been addressed


[DFP-1165]:
https://forbole.atlassian.net/browse/DFP-1165?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ