fix: crash when manually refreshing posts on home screen (#509)
## Description

This PR fixes a crash that occurs when manually refreshing the posts on
the home page.

The origin of the crash was due to the `GetRelationshipForAddress` which
queries for a `user_relationship` type object, which is then used to
check if the user has a relationship with the query subject. However,
the query in question only returns the `creator_address` as a string,
rather than a `RelationshipFields` object, which results in the apollo
cache to throw an undefined error when attempting to generate a key (see
`useClient:48`).

Thus, the fix was to simply replace the queried field with a
`RelationshipFields` object. This should have no effect on the existing
logic as only the length of the data result of the
`GetRelationshioForAddress` query is checked to see if the current user
has an existing relationship with the user-being-queried.

---

### Author Checklist

*All items are required. Please add a note to the item if the item is
not applicable and
please add links to any relevant follow up issues.*

- [x] included the correct [type
prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json)
in the PR title
- [x] provided a link to the relevant issue or specification
- [x] reviewed "Files changed" and left comments if necessary
- [x] confirmed all CI checks have passed

### Reviewers Checklist

*All items are required. Please add a note if the item is not applicable
and please add
your handle next to the items reviewed if you only reviewed selected
items.*

- [ ] confirmed the correct [type
prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json)
in the PR title
- [ ] confirmed all author checklist items have been addressed