refactor: improved post comments visualization speed  (#670)
## Description

Closes: #XXXX

<!-- Add a description of the changes that this PR introduces and the
files that
are the most critical to review. -->

This PR reduces the lag experienced while opening a post comment. The
lag was caused by the query that was computing the difference of
comments from the information in the server and the application state.
To fix this, the query has been removed, and we use the `comments_count`
field of the `Post` object instead of querying it.

---

### Author Checklist

*All items are required. Please add a note to the item if the item is
not applicable and
please add links to any relevant follow up issues.*

- [x] included the correct [type
prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json)
in the PR title
- [ ] provided a link to the relevant issue or specification
- [x] reviewed "Files changed" and left comments if necessary
- [x] confirmed all CI checks have passed

### Reviewers Checklist

*All items are required. Please add a note if the item is not applicable
and please add
your handle next to the items reviewed if you only reviewed selected
items.*

- [ ] confirmed the correct [type
prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json)
in the PR title
- [ ] confirmed all author checklist items have been addressed


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit


- **Refactor**
- Simplified the `PostCard` component by removing the `fetchTimestamp`
property and related logic.
- Updated and streamlined the handling of comments count in multiple
components and hooks.
- Simplified the logic for setting comments count in the
`PostCardBottomBar` component.
- Removed unnecessary variables and props associated with fetching
timestamps.
- Modified GraphQL query fragments to include additional fields like
`likes_count` and `comments_count.
	- Added a new `commentsCount` property to the `Post` interface.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->